requirements.txt
psycopg2-binary
postgres

python-dotenv~=0.18.0
pyspark~=3.2.1

boto3~=1.24.35
